requirements.txt
asgiref==3.2.7
astroid==2.3.3
cffi==1.14.0
colorama==0.4.3
cryptography==2.9
Django==3.0.5
django-cors-headers==3.2.1
django-filter==2.2.0
django-rest-knox==4.1.0
djangorestframework==3.11.0
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.1
mccabe==0.6.1
pycparser==2.20
pylint==2.4.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
wrapt==1.11.2
